Restructure
Restructure of the whole website.

Updated to work with react, changed it so it is using node instead of npm.

No longer able to user router, instead created custom routes and controllers for each endpoint.

Some pages are not loading the css files as of yet, so that is to be updated.

Connection to the database is successful and information can be taken out from it (haven't tested inserting data yet).